AttributesMap changed to LocalAttributesManagementRequest - credentials and operationType added Tests and AMQP consumers changed. UserManagementController changed to require admin credentials

+++ src/main/java/eu/h2020/symbiote/security/listeners/amqp/consumers/AttributesMapConsumerService.java
@@ -5,23 +5,32 @@ import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.DefaultConsumer;
 import com.rabbitmq.client.Envelope;
-import eu.h2020.symbiote.security.communication.payloads.AttributesMap;
+import eu.h2020.symbiote.security.commons.exceptions.custom.InvalidArgumentsException;
+import eu.h2020.symbiote.security.commons.exceptions.custom.UserManagementException;
+import eu.h2020.symbiote.security.communication.payloads.ErrorResponseContainer;
+import eu.h2020.symbiote.security.communication.payloads.LocalAttributesManagementRequest;
 import eu.h2020.symbiote.security.repositories.LocalUsersAttributesRepository;
 import eu.h2020.symbiote.security.repositories.entities.Attribute;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.springframework.http.HttpStatus;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.Map;
 
 public class AttributesMapConsumerService extends DefaultConsumer {
 
     private static Log log = LogFactory.getLog(AttributesMapConsumerService.class);
-    private LocalUsersAttributesRepository localUsersAttributesRepository;
+    private final LocalUsersAttributesRepository localUsersAttributesRepository;
+    private final String adminUsername;
+    private final String adminPassword;
 
-    public AttributesMapConsumerService(Channel channel, LocalUsersAttributesRepository localUsersAttributesRepository) {
+    public AttributesMapConsumerService(Channel channel, LocalUsersAttributesRepository localUsersAttributesRepository, String adminUsername, String adminPassword) {
         super(channel);
         this.localUsersAttributesRepository = localUsersAttributesRepository;
+        this.adminUsername = adminUsername;
+        this.adminPassword = adminPassword;
     }
 
     /**
@@ -41,28 +50,49 @@ public class AttributesMapConsumerService extends DefaultConsumer {
 
         String message = new String(body, "UTF-8");
         ObjectMapper om = new ObjectMapper();
-        AttributesMap list;
+        LocalAttributesManagementRequest request;
         String response;
 
         if (properties.getReplyTo() != null || properties.getCorrelationId() != null) {
-
             AMQP.BasicProperties replyProps = new AMQP.BasicProperties
                     .Builder()
                     .correlationId(properties.getCorrelationId())
                     .build();
-            list = om.readValue(message, AttributesMap.class);
-            log.debug("[x] Received LocalUsersAttributesMap with: " + list.getAttributes().size() + " attributes");
-            localUsersAttributesRepository.deleteAll();
-            for (Map.Entry<String, String> entry : list.getAttributes().entrySet()) {
-                localUsersAttributesRepository.save(new Attribute(entry.getKey(), entry.getValue()));
-            }
+            try {
+                request = om.readValue(message, LocalAttributesManagementRequest.class);
 
-            response = "true";
-            this.getChannel().basicPublish("", properties.getReplyTo(), replyProps, response.getBytes());
-            log.debug("Revocation Response: sent back");
+
+                if (request.getAdminCredentials() == null)
+                    throw new InvalidArgumentsException();
+                // and if they match the admin credentials from properties
+                if (!request.getAdminCredentials().getUsername().equals(adminUsername)
+                        || !request.getAdminCredentials().getPassword().equals(adminPassword))
+                    throw new UserManagementException(HttpStatus.UNAUTHORIZED);
+
+
+                if (request.getOperationType().equals(LocalAttributesManagementRequest.OperationType.READ)) {
+                    Map<String, String> localAttributes = new HashMap();
+                    for (Attribute attr : localUsersAttributesRepository.findAll()) {
+                        localAttributes.put(attr.getKey(), attr.getValue());
+                    }
+                    response = om.writeValueAsString(localAttributes);
+                } else {
+                    localUsersAttributesRepository.deleteAll();
+                    for (Map.Entry<String, String> entry : request.getAttributes().entrySet()) {
+                        localUsersAttributesRepository.save(new Attribute(entry.getKey(), entry.getValue()));
+                    }
+                    response = "true";
+                }
+                this.getChannel().basicPublish("", properties.getReplyTo(), replyProps, response.getBytes());
+                log.debug("Revocation Response: sent back");
+            } catch (InvalidArgumentsException | UserManagementException e) {
+                response = (new ErrorResponseContainer(e.getErrorMessage(), e.getStatusCode().ordinal())).toJson();
+                this.getChannel().basicPublish("", properties.getReplyTo(), replyProps, response.getBytes());
+            }
         } else {
             log.error("Received RPC message without ReplyTo or CorrelationId properties.");
         }
         this.getChannel().basicAck(envelope.getDeliveryTag(), false);
+
     }
 }
